add file limit

+++ src/server/engine_shard_set.h
@@ -49,7 +49,7 @@ class EngineShard {
 
   // Sets up a new EngineShard in the thread.
   // If update_db_time is true, initializes periodic time update for its db_slice.
-  static void InitThreadLocal(util::ProactorBase* pb, bool update_db_time, uint64_t max_file_size);
+  static void InitThreadLocal(util::ProactorBase* pb, bool update_db_time, size_t max_file_size);
 
   static void DestroyThreadLocal();
 
@@ -332,7 +332,7 @@ class EngineShardSet {
   void TEST_EnableCacheMode();
 
  private:
-  void InitThreadLocal(util::ProactorBase* pb, bool update_db_time, uint64 max_file_size);
+  void InitThreadLocal(util::ProactorBase* pb, bool update_db_time, size_t max_file_size);
 
   util::ProactorPool* pp_;
   std::vector<FiberQueue*> shard_queue_;
